actu

+++ Concesionario/src/CapaLogica/Cliente.java
@@ -6,18 +6,18 @@ import java.util.Scanner;
 public class Cliente extends Persona {
 	static ArrayList<Cliente> clientes = new ArrayList<Cliente>();
 	private String modeloInteres;
-	private int presupuesto;
+	private long presupuesto;
 	private Auto auto;
 
 	public Cliente(String nombre, long cedula, long telefono, String correo, String direccion, String modeloInteres,
-			int presupuesto) {
+			long presupuesto) {
 		super(nombre, cedula, telefono, correo, direccion);
 		this.modeloInteres = modeloInteres;
 		this.presupuesto = presupuesto;
 		Cliente.clientes.add(this);
 	}
 
-	public Cliente(String nombre, long cedula, long telefono, String correo, String modeloInteres, int presupuesto) {
+	public Cliente(String nombre, long cedula, long telefono, String correo, String modeloInteres, long presupuesto) {
 		super(nombre, cedula, telefono, correo);
 		this.modeloInteres = modeloInteres;
 		this.presupuesto = presupuesto;
@@ -28,7 +28,7 @@ public class Cliente extends Persona {
 		return modeloInteres;
 	}
 
-	public int getPresupuesto() {
+	public long getPresupuesto() {
 		return presupuesto;
 	}
 
@@ -36,7 +36,7 @@ public class Cliente extends Persona {
 		this.modeloInteres = modeloInteres;
 	}
 
-	public void setPresupuesto(int presupuesto) {
+	public void setPresupuesto(long presupuesto) {
 		this.presupuesto = presupuesto;
 	}
 	
@@ -82,7 +82,7 @@ public class Cliente extends Persona {
 				String modelo = sc.nextLine(); 
 			
 				System.out.print("Introduzca su Presupuesto: ");
-				int presupuesto = sc.nextInt(); 
+				long presupuesto = sc.nextInt(); 
 				sc.nextLine();
 				Cliente cli = new Cliente(nombre, cedula1, telefono, correo, direccion, modelo, presupuesto);
 				System.out.print(cli.toString());
